backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.31
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
numpy>=2.1.0
scipy>=1.14.1
pandas>=2.2.3
duckdb>=1.1.0
langchain-google-genai>=0.2.0
scikit-learn>=1.5.0
shap>=0.51.0
bertopic>=0.17.0
lifelines>=0.30.0
streamlit>=1.56.0
google-genai>=0.3.0
